added my picture and made it fluid (after a lot of experimentation)
I also tried and tried to figure out a simple way to make the nav stay fixed at the top while ALSO making the rest of the page follow nicely with a mobile OR desktop screen-size.
I know that won't be too hard once I start incorporating some JavaScript and JQuery and maybe some other tools... Looking forward to that.